# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kayak rental platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /availability/check:
    post:
      operationId: checkAvailability
      tags: [availability]
      summary: Check product availability for date range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRequest"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResult"
  /pricing/calculate:
    post:
      operationId: calculatePrice
      tags: [pricing]
      summary: Calculate rental price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingRequest"
      responses:
        "200":
          description: Price calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingResult"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderConfirmation"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID (public, for confirmation page)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderConfirmation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AvailabilityRequest:
      type: object
      properties:
        productId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        quantity:
          type: integer
      required:
        - productId
        - startDate
        - endDate
        - quantity
    AvailabilityResult:
      type: object
      properties:
        available:
          type: boolean
        availableQuantity:
          type: integer
        requestedQuantity:
          type: integer
        blockedDates:
          type: array
          items:
            type: string
        message:
          type: ["string", "null"]
      required:
        - available
        - availableQuantity
        - requestedQuantity
        - blockedDates
    PricingRequest:
      type: object
      properties:
        productId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        quantity:
          type: integer
      required:
        - productId
        - startDate
        - endDate
        - quantity
    PricingResult:
      type: object
      properties:
        exactPrice:
          type: ["number", "null"]
        approximatePrice:
          type: ["number", "null"]
        managerReviewRequired:
          type: boolean
        pricingComment:
          type: ["string", "null"]
        tariffUsed:
          type: ["string", "null"]
        days:
          type: integer
        pricePerDay:
          type: ["number", "null"]
        depositAmount:
          type: ["number", "null"]
      required:
        - managerReviewRequired
        - days
    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        startDate:
          type: string
        endDate:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        communicationChannel:
          type: string
          enum: [telegram, whatsapp, phone, email]
        deliveryType:
          type: string
          enum: [pickup, delivery]
        deliveryAddress:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        privacyAccepted:
          type: boolean
      required:
        - items
        - startDate
        - endDate
        - customerName
        - customerPhone
        - communicationChannel
        - deliveryType
        - privacyAccepted
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        lifejacketSizes:
          type: ["string", "null"]
      required:
        - productId
        - quantity
    OrderConfirmation:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryType:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemConfirmation"
        exactPrice:
          type: ["number", "null"]
        approximatePrice:
          type: ["number", "null"]
        managerReviewRequired:
          type: boolean
        pricingComment:
          type: ["string", "null"]
        totalDeposit:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - startDate
        - endDate
        - customerName
        - customerPhone
        - deliveryType
        - items
        - managerReviewRequired
        - createdAt
    OrderItemConfirmation:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        depositAmount:
          type: ["number", "null"]
      required:
        - productId
        - productName
        - quantity
...